+++ weblate/checks/markup.py
@@ -110,13 +110,16 @@ class BBCodeCheck(TargetCheck):
     check_id = "bbcode"
     name = gettext_lazy("BBCode markup")
     description = gettext_lazy("BBCode in translation does not match source")
+    default_disabled = True
+
+    def __init__(self):
+        super().__init__()
+        self.enable_string = "bbcode-text"
 
     def check_single(self, source: str, target: str, unit: Unit):
         # Parse source
         src_match = BBCODE_MATCH.findall(source)
-        # Any BBCode in source?
-        if not src_match:
-            return False
+
         # Parse target
         tgt_match = BBCODE_MATCH.findall(target)
         if len(src_match) != len(tgt_match):